refactor: 금액 관련 데이터 타입을 Integer로 변경

+++ src/main/java/com/tenten/studybadge/participation/service/StudyChannelParticipationService.java
@@ -158,7 +158,7 @@ public class StudyChannelParticipationService {
     private void recordDeposit(StudyChannel channel, Member member, Long amount) {
         StudyChannelDeposit deposit = StudyChannelDeposit.builder()
                 .depositAt(LocalDateTime.now())
-                .amount(amount)
+                .amount(amount.intValue())
                 .depositStatus(DepositStatus.DEPOSIT)
                 .studyChannel(channel)
                 .member(member)
